# src/SvarOMatic.Backend/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Swagger - Svar-o-Matic",
    "description" : "This is the backend for a solution that allows employees of Miles AS to efficiently match incoming project requests with existing consultant expertise.\n\nSome useful links:\n- [The Svar-o-Matic repository](https://github.com/miles-no/svar-o-matic)",
    "termsOfService" : "http://swagger.io/terms/",
    "contact" : {
      "email" : "drift@miles.no"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.11"
  },
...
  "servers" : [ {
    "url" : "http://localhost:5090"
  } ],
...
  "paths" : {
...
    "/previousResponses/{consultantId}/add" : {
      "post" : {
        "tags" : [ "previous-responses" ],
        "summary" : "Add a previous skjema to a consultant profile",
        "description" : "Add an skjema to the profile that the consultant has previously used in an application.",
        "operationId" : "addPreviousSkjema",
        "parameters" : [ {
          "name" : "consultantId",
          "in" : "path",
          "description" : "Id of the consultant that the addition is related to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "description" : "A previous skjema that the consultant has used",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PreviousResponseRequest"
              }
            },
            "application/xml" : {
              "schema" : {
                "$ref" : "#/components/schemas/PreviousResponseRequest"
              }
            },
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "$ref" : "#/components/schemas/PreviousResponseRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input"
          }
        },
        "security" : [ {
          "svaromatic_docs_auth" : [ "write:docs", "read:docs" ]
        } ]
      }
    },
    "/styleGuide/add" : {
      "post" : {
        "tags" : [ "style-guide" ],
        "summary" : "Add/Replace the style guide used by all users of svar-o-matic",
        "description" : "Add/Replace the style guide used by all users of svar-o-matic",
        "operationId" : "addStyleGuide",
        "requestBody" : {
          "description" : "A Style Guide",
          "content" : {
            "application/octet-stream" : {
              "schema" : {
                "type" : "string",
                "format" : "binary"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input"
          }
        },
        "security" : [ {
          "svaromatic_docs_auth" : [ "write:docs", "read:docs" ]
        } ]
      }
    },
    "/docsForConsultant/{consultantId}/get" : {
      "get" : {
        "tags" : [ "cv", "previous-responses", "style-guide", "consultant" ],
        "summary" : "Finds all CVs, skjemaer and style-guides applicable for a consultant",
        "description" : "Multiple docs (CV, skjema and/or style guide) values can be provided with comma separated strings",
        "operationId" : "findDocsForConsultant",
        "parameters" : [ {
          "name" : "consultantId",
          "in" : "path",
          "description" : "Id of the consultant that the search is limited to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "anyOf" : [ {
                      "$ref" : "#/components/schemas/StyleGuide"
                    }, {
                      "$ref" : "#/components/schemas/CV"
                    }, {
                      "$ref" : "#/components/schemas/PreviousResponse"
                    } ]
                  },
                  "x-content-type" : "application/json"
                }
              },
              "application/xml" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "anyOf" : [ {
                      "$ref" : "#/components/schemas/StyleGuide"
                    }, {
                      "$ref" : "#/components/schemas/CV"
                    }, {
                      "$ref" : "#/components/schemas/PreviousResponse"
                    } ]
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid status value"
          }
        },
        "security" : [ {
          "svaromatic_docs_auth" : [ "write:docs", "read:docs" ]
        } ]
      }
    },
    "/toggleUsage/{consultantId}" : {
      "get" : {
        "tags" : [ "previous-responses", "style-guide" ],
        "summary" : "Toggle the usage of a document for a consultant to generate answers",
        "description" : "Toggle the usage of a document for a consultant to generate answers",
        "operationId" : "toggleDocUsageForConsultant",
        "parameters" : [ {
          "name" : "consultantId",
          "in" : "path",
          "description" : "Id of the consultant that the document status is limited to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "description" : "A previous skjema that the consultant has used",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DocUsageRequest"
              }
            },
            "application/xml" : {
              "schema" : {
                "$ref" : "#/components/schemas/DocUsageRequest"
              }
            },
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "$ref" : "#/components/schemas/DocUsageRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "Document not found"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        }, {
          "svaromatic_docs_auth" : [ "write:docs", "read:docs" ]
        } ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "StyleGuide" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "string",
            "format" : "binary"
          },
          "modified" : {
            "type" : "string",
            "format" : "date-time"
          },
          "isInUse" : {
            "type" : "boolean"
          }
        }
      },
      "CV" : {
        "type" : "object",
        "properties" : {
          "Id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "content" : {
            "type" : "string",
            "format" : "binary"
          },
          "owner" : {
            "type" : "integer",
            "format" : "int64"
          },
          "modified" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PreviousResponse" : {
        "type" : "object",
        "properties" : {
          "Id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "content" : {
            "type" : "string",
            "format" : "binary"
          },
          "owner" : {
            "type" : "integer",
            "format" : "int64"
          },
          "originalSubmitter" : {
            "type" : "integer",
            "format" : "int64"
          },
          "submittedOn" : {
            "type" : "string",
            "format" : "date-time"
          },
          "submittedToKunde" : {
            "type" : "string"
          },
          "ledToInterview" : {
            "type" : "boolean"
          },
          "ledToOppdrag" : {
            "type" : "boolean"
          },
          "isInUse" : {
            "type" : "boolean"
          }
        }
      },
      "PreviousResponseRequest" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "string",
            "format" : "binary"
          },
          "originalSubmitter" : {
            "type" : "integer",
            "format" : "int64"
          },
          "submittedOn" : {
            "type" : "string",
            "format" : "date-time"
          },
          "submittedToKunde" : {
            "type" : "string"
          },
          "ledToInterview" : {
            "type" : "boolean"
          },
          "ledToOppdrag" : {
            "type" : "boolean"
          }
        },
        "xml" : {
          "name" : "##default"
        }
      },
      "DocUsageRequest" : {
        "type" : "object",
        "properties" : {
          "Id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "type" : {
            "type" : "string"
          },
          "isInUse" : {
            "type" : "boolean"
          }
        },
        "xml" : {
          "name" : "##default"
        }
      },
...
      "ApiResponse" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "type" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "example" : {
          "code" : 0,
          "type" : "type",
          "message" : "message"
        },
        "xml" : {
          "name" : "##default"
        }
      }
    },
...
    "securitySchemes" : {
      "svaromatic_docs_auth" : {
        "type" : "oauth2",
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "https://coming-soon.swagger.io/oauth/authorize",
            "scopes" : {
              "write:cvs" : "modify docs for a consultant",
              "read:cvs" : "read your docs"
            }
          }
        }
      },
      "svaromatic_inquiry_auth" : {
        "type" : "oauth2",
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "https://coming-soon.swagger.io/oauth/authorize",
            "scopes" : {
              "write:cvs" : "modify inquiries for a consultant",
              "read:cvs" : "read inquiries"
            }
          }
        }
      },
      "api_key" : {
        "type" : "apiKey",
        "name" : "api_key",
        "in" : "header"
      }
    }
  }
}